Timeout for job runs so that we don't keep dead runs forever
In our current state if we have a job that doesn't return a relay entry
it hangs around in state "running" forever.  This is resulting in job
containers piling up.  Here we add a killswitch for job containers
that's set to 10 min.  This **should** be plenty of time for a well run
relay request to return an entry.