Consolidate keep-client standard peers into a single deployment
Keep-clients require unique eth accounts.  When we started doing Kube
deployments we handled this caveman style, issuing a deployment and
configmap assigned to a single keep-client, mapping to a unique eth
account.  Today we use a Kube InitContainer to dynamically provision ETH
resources for the keep-client.  This allows us to squash the
standard-peer deployments and get rid of all the copy/pasted yaml.
Bootstrap peers remain a separate animal as they need to use
pre-configured accounts.