Sign and VerifySignature functions removed from operator package
Signing and signature verification are chain-specific operations
and we decided they should be placed in the chain API. For this
reason, we removed Sign and VerifySignature functions from the
operator package and switched the code using them to call chain
Signing API instead.

operator.Signature and operator.SignatureSize have been removed
as well. The size of the signature depends entirely on the chain.
Signature type alias for []byte does not seem to be really needed
- it does not improve clarity but causes import cycle problems.
For this reason, I decided to use []byte for signatures - this is
consistent with the chain Signing API.

Since chain Signing API exposes PublicKey() []byte, we can use
[]byte in DKGResultHashSignatureMessage. It's better because we are
not bound to ECDSA signatures - there is no guarantee the underlying
chain implementation uses ECDSA signatures.