Provide directory to hold config files needed by Terraform configs
Due to the way the bitcoin helm chart is structured we need to be able
to pass a configured helm values yaml to the bitcoin helm release.
Configured in this context is the list of key/value configuration
options we want present in the bitcoin.conf file.  This potential list
is yugeeeeee so I've limited it to the values we're using, instead of
templating stubs for all the available options.