Eliminated signing ID from random beacon operator contract
Since there is only one relay entry processed at the moment it means we
do not have to mark it with a unique ID - the fact there is only one
request processed in time makes it unique by definition.

This change removes signing ID and signing request counter from the
operator contract.

Client's Go code will be adjusted in a separate commit.