Eliminated relay entry SigningID from Go client
In the previous commit we removed signing ID from the operator contract.
We can have just one ongoing relay entry so we do not need additional
identifier.

Two changes require more detailed explanation:

1. In local.go, I removed the map of relay entries altogether. The
validation we had was not up to date - only one group member submits the
entry, it does not make sense to validate if all signatures on entries are
the same because multiple entry submissions from the same group are not
going to happen - this is validated in other tests.

2. In cmd/relay.go I removed the part capturing request ID. When we run
relay command we are bound to one service contract and one operator
contract. Since there can be only one relay entry processed, the fact
that the relay request was accepted is enough to assume that the relay
entry that is going to be produced is "ours".