Do not pass group public key when submitting relay entry
Public key of the group who is responsible for producing the current
relay entry is saved in the operator contract. We were comparing this
saved version against the value passed by clients. Instead of comparing
those two values, let's just take the value saved in the client and
validate the signature against it.