Do not pass previous entry and seed values when submitting relay entry
We had a security bug - group could collude to produce desired random
number because it could freely choose the data to be signed. Instead of
taking the previous entry value, the group could choose anything and
submit it to relayEntry function.

We fixed it here. New group signature has to be done on the previous
entry value and seed and those value are saved in the contract when new relay
entry is requested.

I had to modify some of our tests because they were also exploiting this
bug and instead of submitting signature on the previous entry, they were
submitting signature on something else.

This is a fix to Solidity, Go client will be modified separately.