Experimental sign function to fix signature format in tests
We need to alter the signature returned from web3.eth.sign to make sure
it passes OpenZeppelin validation code. Adding 27 should fix the problem
with `v` value.

However, it seems there are other problems with the format of signature
from `web3.eth.sign` such that it can be rejected on the
non-malleability check. We need to investigate it further but we decided
to capture the current investigation state with this commit.