Use previous beacon value as a seed to new group selection
The group selection protocol should take the previous beacon value along
with staker-specific values to create a ticket.
The previous implementation passed a seed instead of a beacon value to
group selection.

Another problem was that the seed was used as a G1HashToPoint input to
create H parameter for Pedersen commitment in newProtocolParameters of
member. This is dangerous because an adversary could use a seed of their
choice knowing the discrete logarithm of that seed and as a result, know
the trapdoor to the commitment scheme.

Both issues have been fixed in this commit. I have eliminated the second
parameter (seed) from group selection functions and did some renames to
avoid confusion about what value the parameter represents.