On-chain genesis process refactoring
Our previous implementation of genesis process was based on submitting a
new relay entry signing some predefined input data and as a result, new
group selection had been triggered.

This has a disadvantage that we need to register the expected signature
input data when no group is yet available. This is possible only because
we identify group by public key in the operator contract. If we were to
point to the group responsible for signing the entry by index of that
group, this would not be possible.

However, to implement group termination we need to change how to refer
to the group responsible for signing the entry and it needs to be done
via group index. To do that, we decided to refactor genesis process.

Our new approach assumes the existence of `genesis` function on the
operator contract which can be called only when there are no groups. We
use predefined seed data for the first group selection and for the seed
beacon value but this is exactly the same as with the previous approach
- the signature we were submitting as the first relay entry was always
the same.

The difference is that we can now call `genesis` later if the number of
groups falls to zero for some reason. This may not be the best ultimate
solution but is a desired feature for early testnet when a lot of things
can go wrong and we'd like to be able to "restart" the beacon without
involving all external parties.