Identify currently selected group by its index
We keep the information about which group has been selected to sign the
current relay entry in the SigningRequest structure. So far, we kept
this information as a public key of the selected group. This is pretty
handy but we can't use this approach when we will terminate currently
selected group in case of a relay entry timeout. Going through the
entire array of groups and looking the one with the right public key is
not a good option in Solidity.