Introduced function to report relay entry timeout
When the group does not provide relay entry on time, it should be
terminated and entire signing process (group selection + signing) should
start from the beginning with another group.

Here we restart the process and add group to the list of terminated
groups but we do not update group selection process to do not take the
terminated group into account. This will be done in another commit.