Pass service contract address to signRelayEntry
In case of a relay entry timeout, currently operating group is
terminated and a new group is selected to sign. Relay entry timeout is
reported by nodes from the network who do not participate in the
currently operating group. `msg.sender` is for `reportRelayEntryTimeout`
is the node's address. Instead of using `msg.sender` as the address of a
service contract (this will fail), we need to get the address of the
service contract from the original, timed-out, relay request.