Change accused member behaviour in phase 5.
Until now an accused member didn't participate
in the dispute. Because of that, he didn't mark
the accuser member as DQ if needed and diverged
with the final result from the other members.
In order to avoid such behaviour, it is
better to make an assumption that each member
consider itself as a honest participant and
immediately marks all his accusers as DQ.