Migrate keep-client standard peer deployment to StatefulSet
We found a bug in the current deployment where all 5 keep-client
standard peers can share the same config file.  This is due to each
Deployment replica sharing the same persistent volume.  To maintain a
single config file for each standard peer we've migrated to StatefulSet.
This allows us to use a `volumeClaimTemplates` entry and produce 5
unique persistent volume claims for each standard peer replica, this
results in the behavior we want, where each config produced by an
InitContainer run is persisted independently, ensuring a unique config
for each client.