Reorganize pricing RFC and improve specification precision
The processing of fees is moved to the entry submission stage where rewards and
refunds are paid out.

Request data that needs storing is made explicit.

Method for selecting signing group and retrying in case of timeout is made explicit.

Allow and accommodate no-callback requests.

Add footnotes, including a stochastic DKG triggering scheme to buffer the effect
of demand fluctuations on the number of signing groups.