Refactored how notifications are set up
Instead of having setupNotifications function we now have buildNotifiee
which accepts no parameters and returns libp2pnetwork.Notifiee
interface. That interface is later installed into libp2p Network notify
system.

That enforces the responsibility of Connect as the final assembler of the
libp2p setup, while leaving most of the boilerplate code in this function.