Fixed relay entry signing threshold value
We were incorrectly increasing threshold by one. For a group size of 5
members and threshold 3, three honest members were considered not enough
to produce a signature. This was incorrect. Everywhere in our code
'threshold' represents an honest threshold - minimum number of honest
members to complete the protocol successfully.

We will tweak the documentation to 'threshold' value in a separate
commit.