Create keep-test-contract-data bucket
Up to this point we've been doing some going around or ass to find our
elbow on a couple tasks related to private testnet participant
deployment bundling.  One of these tasks is around staking participant
accounts.  When we stake we need the most recently migrated contract
json for KeepToken and TokenStaking.  We've been getting these files by
pulling the latest keep-client provisioner InitContainer image locally,
starting it, and docker cp'ing the contract files out.  Now that we're
moving to script the bundling of client deployments we need something a
bit less crazy.  We'll use circle to publish contract data to a bucket
after each migration.  Here we create that bucket.