Add step for dir creation
The directory structure for bundling deployments has changed a bit.  The
relevant part is that we're sourcing deployment bundle files from
several different locations, instead of all files being managed manually
under one `keep-client-deployment-bundle` dir per participant.  As a
result we can't tar up a nice neat prepared dir for each participant.
This change is small in terms of effort on the participant side, and
allows us to source generic files for tar'ing that are the same for each
participant.