Dropped protocol_result_test.go
This test is always passing and is testing nothing.

In the past, we were returning a failure from DKG in case we did not
meet the required threshold. Currently, we try to complete DKG no matter
what. This is not a final solution and we'll have to refine it to fail
when needed but `protocol_result_test.go` does not make any sense for
the time being. Test is always passing and even if we fix it, it will
cover logic no longer existing.