Readme clarifications
In script execution we cut out the bit about ordering, while it's
possible to execute in different orders, we're leaving it as described
to keep things simple.  We removed the directory/file counts from tree
output as it's not really providing value.  Lastly we clarified the
working directory for running scripts.