Integration test - DKG phase 5 - unfounded accusation
Here we implement a new integration tests for phase 5. This test
scenario requires an access to the internals of the accusing member as
we need to reveal private key established between accuser and accused.

We implemented a helper tool used to hijack communication between two
members and replace messages send in the first three phases of DKG
protocol. With this tool, we can access internals of member and
implement integration tests for scenarios when access to symmetric key
is required.