By default, man-in-the-middle does not accuse anyone
By default, man-in-the-middle does not perform any accusations and do not
resolve them. All accusation are dropped by default. This behavior may be
overwritten by test.

This behavior is easier to write integration tests with since we do not
have to repeat the same "drop-accusation" code over and over again in
our tests.