Set announce addresses env vars
We have two new Circle context provided env var values that we need to
set in the initcontainer docker image.  These are used to provide
non-default annouce addresses for bootstrap and standard peers.  These
values are fed downstream to the keep-client provisioner script to set
the values in the keep-client config files.