Add LibP2P.AnnoucedAddresses to bootstrap peer
Here we provide the config key to setup non-default announce addresses
for a keep-client of type bootstrap.  Leaving this empty, as set in the
config template will use the default value for for announce address.
This value should be overridden when we have a know IP/DNS entry or
entries we want to channel traffic through.