Add Circle CI test results for contract tests
This is handled by adjusting the mocha test runner in Circle CI to use
the mocha-junit-reporter. Currently this requires some jq voodoo, though
in the future it could use a multi-reporter setup or similar and mocha
6.x to reference the built-in xunit reporter.

npm install now runs during a setup step rather than as part of the test
execution command.