Fixed ticket sort algorithm
Tickets should be sorted by their value and tickets with the lowest
value should be submitted first as they have the highest chance of being
selected to the group.

Our previous implementation of sorting algorithm was buggy - instead of
sorting by value it was sorting by virtual staker index. Additionally,
it was replacing just the virtual staker index in the proof instead of
swapping tickets in slice during the stable sort.