Migrate bootstrap peer Deployment to StatefulSet
Using a StatefulSet for bootstrap peers will future proof us when we're
able to scale beyond 1 replica per Deployment.  For this to happen we'll
need to be able to sort out how to apply arbitrary ETH accounts to a
bootstrap peer, rather than a known address that's pre-configured.
Additionally we're introducing a persistent volume for the path on
Storage.DataDir.  It's idomatic to use a StatefulSet for applications
who need to persist state beyond container cycling.