Use multiple volumeClaimTemplates
Combining PersistentVolumeCliams with StatefulSets can have some pretty
gross results. The jist here is that PersistentVolumeClaims have no
assignment until something reuqests the resource.  If a replica from a
StatefulSet requests the resource and their in different AZ's you have
an affinity problem.  If a replica from a StatefulSet with > 1 replica
requests a PersistentVolumeClaim that has ReadWriteOnce permission and
the Claim is already mounted, you get stuck because the second replica
requesting can't mount the volume.  volumeClaimTemplates address this by
acting as a PersistentVolumeClaim controller, assigning a claim or
claims to a replica with unique assignment.