Factor keep-* Service configurations to own files
I'm not a huge fan of lumping all Kube object configurations for a
"thing" into a single file. "think" in this case would be something like
a keep-client-standard peer.  Trying a different approach in having
configuration files represent a single object or multiple of the same
object type and that filename reflect the object type.  We'll see how it
goes.