Simplified the calculation of delay factor and rewards
Before we return delay factor value we divide it by decimals multiplier.

The reason for adding decimals multiplier is to be able to perform float
division on integers. Since delay factor is calculated as a square,
decimals multiplier is also squared. To avoid leaking the square out of the
getDelayFactor, we divide the result by decimals and return the delay
factor as a number with 16 decimals, just like the documentation says.