Workaround for go-ethereum problem with gas estimation
When estimating gas price in `go-ethereum` ABI, `tx.gasprice` is not
passed to the function doing the estimation. This leads to incorrect gas
estimates when code execution path depends on `tx.gasprice`.

This was fixed today in
https://github.com/ethereum/go-ethereum/pull/20189

We need to make a workaround in the contract code before this fix will
be released.