Take the minimum of tx.gasprice/price feed for callback reimbursement
At the time of the request, we add a fluctuation margin. Miner-submitter
may set tx.gasprice to the highest possible value to always get the
entire amount with estimated gas price at the time of the request even
though the gas price dropped or it did not rise to the level justifying
burning the fluctuation margin.