Clarify automatic vs. triggered checks for failures
Some failures can be verified without additional proof,
using only on-chain information,
but require a transaction to trigger this check.

For failure to finish DKG the trigger for another DKG can be used.

Failure to produce an entry requires its own trigger transaction
to penalize and terminate the failed group
and to hand entry generation off to the next one.
The submitter of this transaction needs to be rewarded for it,
but the reward has to be limited to avoid certain attacks.