Remove unused ETH Docker configs
When we started experimenting with ETH builds we had a few different
images.  geth-node-with-dag was an attempt to try and save some time on
boot by having DAG already baked into the image.  geth-dev was fine, we
jsut don't need it anymore as the geth-node build is using the same
image.