Don't stake if client type is relay-requester
In our private testnet we're seeing a repeating issue where after a few
days of operation, our relay-requester starts failing to stake accounts.
This results in the requester to run in an InitContainer loop, creating
and trying to stake accounts.  The crux of this is that at our current
staking amount (per account), our token contract has enough tokens to
stake 1000 times.  After a few days of operation we hit that number as
re request a relay entry every 5 minutes.  After dicussing the issue
Piotr pointed out that we don't need to stake relay-requster clients at
all.  They're consumers of the network, not operators.  Here we add some
additional logic to prevent staking when keep-client-type is
relay-requester.  Additionally I moved the staking decision logic inside
the staking function, it felt like a better fit there.