Include reference for Instanbul block
Frankly, I'm not sure if this is needed or not.  Had a look at the
go-ethereum code and a value of 0 indicates the chain is already on the
fork.  Seeing as how we're building fresh chains I would expect the EIPs
for some fork to be included on the fresh chain so long as we use the
geth client version that includes all of those EIPs.  Anywho, adding it
for documentation sake.