Store group public key on-chain uncompressed
The cost of verification of BLS signature on a pre-Istanbul client was
about `1 177 486` gas and the most expensive part was `g2Decompress`
with a cost of about `831 961`. Going down the execution path, it turned
out that the problem lies in `gfP2Pow` function in our Solidity AltBn128
library with a single invocation of a cost of about `271 919` gas and
having three such invocations for BLS verification.

It seems that - as for today - there is no known efficient way of
decompressing to G2 avoiding expensive looping. Storing group public key
uncompressed on-chain will cost additional `20 000` gas but this is a
cost paid once, during DKG result submission.

In this commit, we update client and contracts to store group public key
in an uncompressed form and save `831 961 - 20 000` gas for each new
relay entry.