Always pull image
The cluster caches images.  If there's a deployment and the image tag
hasn't changed, a fresh image isn't pulled.  We're tagging the ETH
deployments with the geth version.  We may tweak configurations without
changing the ETH version, thus publishing with the same tag.  We need to
tell the cluster to always pull, to ensure changes are grabbed.