Use a single keep-client-config template
As we converge on a more straight forward setup we've introduced a
keep-client config template that can be used for both "bootstrap" and
"standard" peers.  The caveat to this is that the ETH account to be used
by some "bootstrap" peer will need to be defined in the kube config for
that peer.  This is done with an env var KEEP_CLIENT_ETH_ACCOUNT in the
provision-keep-client InitContainer envs section.  The end result of
this is that we can setup as many bootstrap deployments as well like
using the pre-generated ETH accounts on our ETH network.  Along with
setting the ^ env var, there needs to be a config map with the
associated keyfile for us to download.  This is a temporary measure
until we can move to something more dynamic.