Adding new configurable options
We've migrated to a new universal config template.  As part of this
there is a new option we need to configure on a bootstrap peer Kube
config.  The ETH account is now set on env var
KEEP_CLIENT_ETH_ACCOUNT.  This will be passed to the config and combined
with the associated Kube ConfigMap to supply the keep-client with ETH
account material.