Hardcoding bootstrap peers in config template
Having configurable bootstrap peers turns out to be a bit more
complicated than I expected.  With the way the env var value for peers
gets parsed in tomlfy, we run into issues with > 1 peer configured, as
it's captured as one large string.  I can probably work with this using
a custom parser in the provisioner script but I've got too many other
fish to fry at the moment.  What matters is multi-bootstrap and we're
providing that here.