Create local copy of transactor options for modifications
We were creating a new copy of transactor options only when the function
was payable. As a result, any modification of either GasLimit or
GasPrice was propagated to the contract-level transactor options and
could make another function call fail.

What is more, the way we were creating a local copy of transactor
options was wrong - instead of creating a new instance of the structure
we were creating a new instance of a pointer with the same value as the
original one.

The code was modified to always create a new copy of transactor options
and the way we create a copy was fixed.