Reference ethereum config struct from keep-common
The config struct now lives in keep-common so that a common
understanding of Ethereum configuration can be shared by Keep
applications that interact with it. Note that this does not prescribe
*where* in an application's configuration the Ethereum configuration
will be defined; that is still left up to the application, and in
keep-core's case the structure remains unchanged.