Add config.ReadEthereumConfig helper
This exported helper enables reading just the Ethereum portion of a
config directly, without having to worry about the structure of the
surrounding config. This will be helpful when keep-common-generated code
needs to read Ethereum configuration and act on it, without being
interested in which application is using that generated code and what
config structure that particular application has chosen to use. All that
matters in this case is that the Ethereum configuration's structure is
as expected.