Replacing mapping with array to simulate a linked list
- Replaced `previousTicketIndex` map (uint256 -> uint256) with the uint8[] array.
Indices in the array mimic a `key` part and the values mimic the values
in this map. It was done to reduce gas fees from reading the map.
"Uint8 array of 64 members can fit in just 2 storage fields, so we can read the
entire array for just 2 * SREAD = 1600 gas instead of paying up to
64 * SREAD = 51,200 gas for a mapping(uint256 -> uint256)".
- Changed uint256 to uint8 in a couple of places.