Do not save ticket proof during submission
The previous version of group selection protocol included off-chain
validation of submitted tickets and possible complaints. For that reason,
proofs had to be stored in the contract to provide an information for
complaint resolution.

The recent version of group selection protocol does not contain
off-chain tickets validation. All tickets are validated on-chain during
their submission. For this reason we do not need to store the full proof
on-chain once it has been validated.

This change brings down gas cost for ticket submission quite
significantly.