Protobuf network message retransmission header changes
1. Renamed `retransmissionSequence` to `retransmission` and changed the
type from int32 to uint32.

It is not possible retransmission number can be negative so uint32 is
better.

2. Removed checksum field.

We do not want to trust the sender whether the checksum is correct and
it reflects the message type and payload. We'd have to validate the
checksum on the receiver side anyway. If the receiver needs to
validate the checksum, there is no need to transfer it in the message
from sender.