Initial libp2p retransmitter implementation
libp2p pubsub does not guarantee message delivery. Messages can be
dropped on multiple levels, sometimes even before they leave the
sender's computer.

To increase the chance of message delivery we need to build something on
a top of the pubsub. This is what we are hoping to achieve with the
retransmitter introduced here.

For each phase when we need to communicate with other members, messaging
takes 3 blocks. From all that time, we are active for the first few seconds
when the message exchange happens. The rest of the time we are idle.
Instead of being idle, we can automatically retransmit our messages.

The retransmitter should automatically retransmit messages and filter
out retransmitions so that message is guaranteed to be received no more
than one time.