Upgrade go-ethereum dependency to v1.9.7
Here we upgrade go-ethereum to the most recent release. Instead of
using replace instruction to point to our forked version, we now
directly refer to the upstream.

The reasoning behind forking repositories was that nobody can delete
things in our owned repositories. However, there is still a problem
with transitive dependencies. We decided neither of these things
should be a blocker and that we are fine with sticking to upstream
for now and figuring out what we want to do about pinning in
December/January.

We also upgrade keep-common dependency to the one with upgraded
go-ethereum version.