Support for overloaded payable functions
Solidity allows for function overloading and Go does not. Previous
versions of go-ethereum had a bug that only one from the overloaded
functions had Go binding generated.

We were lucky, because the requestRelayEntry function variant we use
in utility.go is declared as a second variant in the contract and
because of that, it was the one generated to Go. Because of go-ethereum
bug, the first variant was completely ignored but we were not using it
from Go so build was passing.

go-ethereum fixed this problem in version 1.9.0
(https://github.com/ethereum/go-ethereum/pull/17099) and for
overloaded function, variants are named: function, function0,
function1...

In this commit we adjust our code detecting if the function is payable
to follow the function variants naming strategy from go-ethereum ABI
generator.