Fixed abigen installation command
Two changes related to abigen command installation:

1. Added 'apk add linux-headers'

Without that, 'github.com/karalabe/usb' which is a dependency of
'abigen' fails compilation with:

../../../../karalabe/usb@v0.0.0-20190919080040-51dc0efba356/libusb
/libusb/os/linux_usbfs.h:24:10: fatal error: linux/types.h: No such
file or directory

2. Changed 'go install .' to 'go get' for 'abigen'

'go-ethereum' 1.9.7 is still on govendor. When go modules tries to
resolve dependencies of abigen from 'vendor.json' they download wrong
version of protobuf (1.2.x instead of 1.3.x) and the compilation fails
with undefined: 'proto.ProtoPackageIsVersion3' error.

This is probably caused by the fact go modules does not well tolerate
vendored, non-tagged dependencies and 'vendor.json' contains a reference
to non-tagged protobuf commit.

This solution is a temporary workaround. We hope to switch back to 'go
install .' after 'go-ethereum' migrates to go modules which should
happen in version 1.9.8.