withdrawFromGroup function made public
Since we use DELEGATECALL for a library function call, we preserve
msg.sender for a public function. We can then make this function public
and use msg.sender instead of having it internal and passing msg.sender
as an argument from the operator contract. This will let us save some
operator contract size as the function will no longer be copied to the
operator contract bytecode.