Additional expired/stale group assertions
Adding those assertions as a guarantee that our tests are always
executed for a valid input data as it is quite hard to simulate
expired/stale groups and it is better to double-check than build our
assumptions about the state of the code on false positives.

One removed assertion was redundant - if group is not expired it can't
be stale.