Revert group index search for non-existing groups
When group could not be found we should revert the transaction instead
of returning 0 since 0 is a valid index of a genesis group.

The same reasoning has been applied to isExpiredGroup in the stub
contract. false is a valid return value so it is better to revert when
the group could not be found.