Increased block gas limit to current Ropsten value
The previous gas limit 0x47B760 = 4 700 000 is about 50% of the current
gas limit on Ropsten which is 0x7A1200 = 8 000 000.

Truffle increased the default gas limit from 4 712 388 to 6 721 975
and the previous gas limit set in genesis block did no longer allow us
to deploy contracts.

Here we set the gas limit to the one used currently on Ropsten. Please
note that the current gas limit on Mainnet is even higher: ~10 000 000.