Implement slashing for relay entry timeout
When a group fails to produce an entry, all of its members shall be
subject to seizing. The submitter of the trigger transaction shall be
treated as the tattletale, but the tattletale reward shall be limited to
min(1, 20 / group_size) of the maximum, or effectively the minimum stake
of a single member. This is to prevent actors in a lynchpin position from
profitably stealing other stakers' funds.