Separate entry generated and entry submitted events
Threshold relay state can be watched from two perspectives: customer and
operator.

Customer wants to request for a new relay entry and be informed when new
relay entry arrives. Customer cares about the value of the generated
entry.

Operators wants to submit a new relay entry and be informed when new
relay entry has been submitted. Operator does not care about the value
of the generated entry because it either knows that value (by
participating in the signing group) or does not care about this value
(by not participating in the signing group and only monitoring entry
submission).

For those two reasons we have two events fired on chain - one in the
operator contract and one in the service contract. In this commit I've
made this separation clear in the off-chain client code. That allowed to
remove all the parameters from SignatureSubmitted event in the operator
contract and reduce gas cost of entry verification a little.

What is more important, this change will help to refactor how previous
entry is stored on the chain and eliminate expensive g1Decompress and
g1HashToPoint operations during BLS signature verification.

Relay entry verification gas changes obtained from
TestKeepRandomBeaconOperatorRelayEntry.js using non-Istanbul client:

estimate: 448828 -> 448544 (-284)
gasUsed: 433828 -> 433544  (-284)