Eliminate relay request seed from on-chain contract code
Allowing customer to provide seed is not secure in the current model.
The attacks that use requster-provided seeds are fundamentally
brute-force but that does not make them unviable. The attacker does not
have to e.g. break encryption of 128-bit keys but even a 1-in-1.000.000
outcome in biasing group selection can be useful for the attacker.

This change will also let us to avoid expensive g1HashToPoint operation
during relay entry verification as the input of the beacon will be
always G1.