Use G1 message for BLS verification and do not compress signature
With this change we remove the need of executing g1Decompress and
g1HashToPoint functions inside BLS.verify what will reduce the gas cost
of BLS verification.

Those two functions were two the most expensive parts of BLS
verification, not including the pairing check.

Since in the threshold relay we'll be always signing the previous value,
and the previous value is G1 point, we can ommit g1HashToPoint call.

Also, if we store signature uncompressed, we can avoid calling
g1Decompress. This will slightly increase the cost of storage but
g1Decompress is much more expensive so it is a win anyway.