Store relay entry as uncompressed as G1 point
Here we adjust the code of operator and service contract to store and
pass relay entry as an uncompressed G1 point. This change is a result of
BLS library refactoring from previous commits.

This will allows us to reduce the cost of relay entry verification by
avoiding expensive g1Decompress and g1HashToPoint calls.