Separate miner service, add network scope
We're running multiple ETH deployments in keep-dev, one to Ropsten and
the other internal.  To distinguish configuration files we're adding a
network type identifier.  Along with this we're separating configs by
Kube object type to try and keep configs easier to maintain.  Here we
migrate the miner to a DaemonSet to try and further reduce chain state
loss incidents.