Remove ConfigMap files
We generate these ConfigMaps from files, via a script.  The intent
behind downloading the configmap here from Kube was to have a checked-in
record of state at any given point.  It's a bit noisy so I'm going to
get rid of it for now.  I'll see about bringing this back when > 1
person is working in this space.