Add support for staking multiple accounts on ropsten
This one is a doozy, sorry future maintainers.  To stake an account on
ropsten we need to an instance of HDWalletProvider in conjunction with
an account address to create a proper signature for use in the
stakeOperatorAccount function.  For now we're maintaining 5 operator
accounts on ropsten, each one needs a profile that includes its address
and HDWalletProvider.  We store this info for each operator in an array,
then store each of those arrays in a map.  To stake we iterate over this
map and run the staking function against each account
address/HDWalletProvider pair.