Fail isStaleGroup check when group could not be found
Instead of returning 'false' and potentially leading client to do some
unexpected storage operations, we should be clear that the group with
the provided public key is not known.

The previous approach caused some unexpected problems when we were
using compressed version of public key to execute the isStaleCheck
function. Having revert for unkown groups would let us realize we have the
problem ealier and what is the nature of the problem.