Use uncompressed group public key evereywhere but in on-disk storage
Instead of compressing the key on-chain to send it in the event and
using compressed key in the in-memory registry in the client, we
now use uncompressed group public key everywhere but in the on-disk
storage.

Groups library holds public key uncompressed, DKG sends public key
uncompressed, relay entry request event contains uncompressed public
key, and in-memory groups registry keeps groups map with uncompressed
public key as map keys.

The only reason on-disk storage uses compressed public key is because
uncompressed versions are too long to use them as directory names.

This version makes relay request TX a bit cheaper.
Before: ~233966, After: ~230952